improve column width calculation algorithm (#2)
This replaces the naive column width calculation with something better.
It first tries to fit all columns using their maximum requested width.
If this is not possible it will try to give each column its fair space
share which may be less than the requested maximum, but will be at least
the requested minimum width. As a last resort it will try to allocate
the requested minimum width where possible, truncating columns where not
even this is possible.